+++ releasenotes.md
@@ -8,7 +8,7 @@
 * Using the `vanilla` CCMD in an alias will now work correctly.
 * The `bind` CCMD can now be used to bind a string of commands to a control. For example, to enable vanilla mode without lowering the graphic detail by pressing the <kbd>V</kbd> key, enter `bind 'v' "vanilla; r_detail high"` in the console.
 * A custom message can now be displayed using the new `print` CCMD.
-* The tossing of items dropped by monsters when they are killed can now be toggled using the new `tossdrop` CVAR. (This CVAR is disabled when vanilla mode is on.)
+* The tossing of items dropped by monsters when they are killed can now be toggled using the new `tossdrop` CVAR. (This CVAR is `off` when vanilla mode is enabled.)
 * If they can be found, the *Final DOOM* IWADs (`plutonia.wad` and `tnt.wad`) will now automatically be loaded for certain PWADs that require them.
 
 ---